+++ rust/bridge/shared/types/src/node/chat.rs
@@ -64,6 +64,25 @@ impl ChatListener for NodeChatListener {
         });
     }
 
+    fn received_alerts(&mut self, alerts: Vec<String>) {
+        let roots_shared = self.roots.clone();
+        self.js_channel.send(move |mut cx| {
+            let callback_object_shared = &roots_shared.callback_object;
+            let callback = callback_object_shared.to_inner(&mut cx);
+            let js_alerts = cx.empty_array();
+            // We use zip instead of enumerate here so that i is a u32 rather than usize.
+            for (alert, i) in alerts.into_iter().zip(0..) {
+                let js_alert = cx
+                    .try_string(alert)
+                    .unwrap_or_else(|_| cx.string("[invalid alert]"));
+                js_alerts.set(&mut cx, i, js_alert)?;
+            }
+            let _result = call_method(&mut cx, callback, "_received_alerts", [js_alerts.upcast()])?;
+            roots_shared.finalize(&mut cx);
+            Ok(())
+        });
+    }
+
     fn connection_interrupted(&mut self, disconnect_cause: DisconnectCause) {
         let disconnect_cause = match disconnect_cause {
             DisconnectCause::LocalDisconnect => None,