+++ rust/net/src/chat/ws2.rs
@@ -28,6 +28,7 @@ use tokio_stream::wrappers::{ReceiverStream, UnboundedReceiverStream};
 use crate::chat::{
     ChatMessageType, ChatServiceError, MessageProto, Request, RequestProto, Response, ResponseProto,
 };
+use crate::env::ALERT_HEADER_NAME;
 use crate::infra::ws::TextOrBinary;
 use crate::infra::ws2::{MessageEvent, NextEventError, TungsteniteSendError};
 
@@ -72,6 +73,12 @@ pub struct Config {
 
 #[derive(Debug)]
 pub enum ListenerEvent {
+    /// Zero or more alerts were received from the server.
+    ///
+    /// These are more lightweight than the full requests of [`Self::ReceivedMessage`].
+    /// They're also not stateful, so "zero alerts" means "clear any previous alerts".
+    ReceivedAlerts(Vec<String>),
+
     /// A request was received from the server.
     ///
     /// The accompanying [`Responder`] can be used to send a response for the
@@ -149,9 +156,10 @@ impl Chat {
     pub fn new<T>(
         tokio_runtime: tokio::runtime::Handle,
         transport: T,
+        connect_response_headers: http::HeaderMap,
         config: Config,
         log_tag: Arc<str>,
-        listener: EventListener,
+        mut listener: EventListener,
     ) -> Self
     where
         T: WebSocketStreamLike + Send + 'static,
@@ -161,6 +169,8 @@ impl Chat {
             local_idle_timeout,
             remote_idle_timeout,
         } = config;
+
+        Self::report_alerts(connect_response_headers, &mut listener);
 
         // Enable access to tokio types like Sleep, but only for the duration of this call.
         let _enable_tokio_types = tokio_runtime.enter();
@@ -178,6 +188,15 @@ impl Chat {
             listener,
             tokio_runtime,
         )
+    }
+
+    fn report_alerts(connect_response_headers: http::HeaderMap, listener: &mut EventListener) {
+        let alerts = connect_response_headers
+            .get_all(ALERT_HEADER_NAME)
+            .iter()
+            .map(|value| value.to_str().unwrap_or("[non-ASCII alert]").to_owned())
+            .collect_vec();
+        listener(ListenerEvent::ReceivedAlerts(alerts))
     }
 
     /// Sends a request to the server and waits for the response.
@@ -1188,6 +1207,7 @@ mod test {
     use http::HeaderMap;
     use test_case::test_case;
     use tokio::select;
+    use tokio::sync::mpsc::error::TryRecvError;
 
     use super::*;
 
@@ -2003,5 +2023,44 @@ mod test {
 
         assert_eq!(listener_rx.recv().await, Some(true));
         assert_matches!(listener_rx.recv().await, None);
+    }
+
+    #[test]
+    fn reports_alerts() {
+        let (listener_tx, mut listener_rx) = mpsc::unbounded_channel();
+        let mut listener_tx: EventListener =
+            Box::new(move |evt| listener_tx.send(evt).expect("can send"));
+
+        Chat::report_alerts(http::HeaderMap::default(), &mut listener_tx);
+        assert_matches!(
+            listener_rx.try_recv().expect("present"),
+            ListenerEvent::ReceivedAlerts(alerts) if alerts.is_empty()
+        );
+        assert_matches!(listener_rx.try_recv(), Err(TryRecvError::Empty));
+
+        Chat::report_alerts(
+            http::HeaderMap::from_iter(
+                [
+                    ("unrelated", "other"),
+                    (ALERT_HEADER_NAME, "first"),
+                    ("yet-another", "something"),
+                    (ALERT_HEADER_NAME, "second"),
+                    ("last-one", "x"),
+                ]
+                .map(|(name, val)| {
+                    (
+                        http::HeaderName::from_static(name),
+                        http::HeaderValue::from_static(val),
+                    )
+                }),
+            ),
+            &mut listener_tx,
+        );
+
+        assert_matches!(
+            listener_rx.try_recv().expect("present"),
+            ListenerEvent::ReceivedAlerts(alerts) if alerts == ["first", "second"]
+        );
+        assert_matches!(listener_rx.try_recv(), Err(TryRecvError::Empty));
     }
 }